Aggiustato un calcolo fatto non troppo bene

+++ Lezione-6/src/main/test_keep.c
@@ -37,7 +37,7 @@ void test(FILE* inp, FILE* out, int n, int m) {
 		printf("Hits %d\n", hits);
 	}
 
-	avg = (double) sumHits/n;
+	avg = (double)sumHits/(double)n;
 	fprintf(out, "%f,%lf\n", (float)n/(float)m, avg);
 	printf(" Hit medie: %f\n", avg);
